Implement the kom i gang route as a sanity based route
